# backend/routes/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.4.0",
    "title": "Sailman project API documentation"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/announcement": {
      "post": {
        "tags": ["announcement"],
        "summary": "Create new announcement",
        "description": "",
        "operationId": "announcement",
        "requestBody": {
          "content": {
            "form-data": {
              "schema": {
                "$ref": "#/components/schemas/Announcement"
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User not logged in"
          }
        }
      }
    },
    "/announcements": {
      "get": {
        "tags": ["announcement"],
        "summary": "Get the list of announcements",
        "description": "",
        "operationId": "announcements",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementInfo"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/category": {
      "post": {
        "tags": ["category"],
        "summary": "Add new category",
        "description": "Requires role: admin",
        "operationId": "category",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/categories": {
      "get": {
        "tags": ["category"],
        "summary": "Get the list of categories",
        "description": "",
        "operationId": "categories",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnnouncementInfo"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "accessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    },
    "schemas": {
...
      "Announcement": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "categories": {
            "type": "object"
          },
          "tags": {
            "type": "object"
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "price": {
            "type": "integer"
          },
          "currency": {
            "type": "string"
          },
          "includedInPrice": {
            "type": "array"
          },
          "notIncludedInPrice": {
            "type": "array"
          },
          "yacht": {
            "type": "string"
          },
          "organizer": {
            "type": "string"
          },
          "lastMinute": {
            "type": "boolean"
          },
          "tidalCruise": {
            "type": "boolean"
          },
          "language": {
            "type": "string"
          },
          "allRoutes": {
            "type": "object"
          },
          "photo": {
            "type": "string",
            "format": "binary"
          },
          "gallery": {
            "type": "object",
            "format": "binary"
          }
        }
      },
      "AnnouncementInfo": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "categories": {
            "type": "object"
          },
          "tags": {
            "type": "object"
          },
          "postedBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
    }
  }
}